add orderbookless treatment
new config option show_order_book allows hiding of the bid/ask/trade boxes.
it also introduces a new 'standalone' order entry box that appears when the order book is hidden.
this shows whatever the price and quantity of the proposed order is, whether it's a bid or ask and has a button to submit it to the exchange.